moving unit tests to their own place
