+++ package.json
@@ -5,7 +5,7 @@
   "author": "Brad Sneade",
   "license": "MIT",
   "scripts": {
-    "test": "jasmine-ts \"src/**/*.test.ts\"",
+    "test": "jasmine-ts \"test/**/*.test.ts\"",
     "build": "tsc",
     "watch": "tsc -w",
     "lint": "tslint -c tslint.json -p tsconfig.json",
@@ -22,6 +22,7 @@
     "karma": "^1.5.0",
     "karma-typescript": "3.0.8",
     "ts-mockito": "^2.2.7",
+    "tslint": "^5.9.1",
     "typescript": "^2.5.3"
   }
 }
